Move the components.
