+++ packages/components/HelpCenterButton.js
@@ -2,9 +2,9 @@ import React from "react";
 import PropTypes from "prop-types";
 import styled from "styled-components";
 
-import { colors } from "@yoast/components/style-guide";
+import { colors } from "./style-guide";
 import { YoastButton } from "./YoastButton";
-import { SvgIcon } from "@yoast/components";
+import SvgIcon from "./SvgIcon";
 
 /**
  * Returns an icon button that can optionally contain text.
