+++ packages/components/SectionTitle.js
@@ -2,8 +2,8 @@ import React from "react";
 import PropTypes from "prop-types";
 import styled from "styled-components";
 
-import ScreenReaderText from "../../../../a11y/ScreenReaderText";
-import { colors } from "@yoast/components/style-guide";
+import ScreenReaderText from "./a11y/ScreenReaderText";
+import { colors } from "./style-guide";
 
 export const StyledTitleContainer = styled.span`
 	flex-grow: 1;
