+++ packages/components/StackedProgressBar.js
