+++ packages/components/Toggle.js
@@ -2,11 +2,11 @@
 import PropTypes from "prop-types";
 import React from "react";
 import styled from "styled-components";
-import { colors } from "@yoast/components/style-guide";
+import { colors } from "./style-guide";
 import { __ } from "@wordpress/i18n";
 
 // Internal dependencies.
-import { getRtlStyle } from "@yoast/components";
+import { getRtlStyle } from "./utils/helpers/styled-components";
 
 const ToggleDiv = styled.div`
 	display: flex;
