+++ packages/components/YoastInput.js
@@ -2,7 +2,7 @@
 import styled from "styled-components";
 
 // Internal dependencies.
-import { colors, rgba } from "@yoast/components/style-guide";
+import { colors, rgba } from "./style-guide";
 
 export const YoastInputContainer = styled.div`
 	display: flex;
