+++ packages/components/YoastLinkButton.js
@@ -1,6 +1,6 @@
 import PropTypes from "prop-types";
 import styled from "styled-components";
-import { colors } from "@yoast/components/style-guide";
+import { colors } from "./style-guide";
 import { addButtonStyles } from "./YoastButton";
 
 /**
