+++ packages/components/YoastModal.js
@@ -3,8 +3,8 @@ import PropTypes from "prop-types";
 import Modal from "react-modal";
 import styled from "styled-components";
 
-import { colors } from "@yoast/components/style-guide";
-import { SvgIcon } from "@yoast/components";
+import { colors } from "./style-guide";
+import SvgIcon from "./SvgIcon";
 
 const StyledHeading = styled.h1`
 	float: left;
