+++ packages/components/composites/SynonymsInput.js
@@ -3,8 +3,8 @@ import PropTypes from "prop-types";
 import uniqueId from "lodash/uniqueId";
 import styled from "styled-components";
 
-import { YoastInputContainer, YoastInputField, YoastInputLabel } from "./YoastInput";
-import { getRtlStyle } from "@yoast/components";
+import { YoastInputContainer, YoastInputField, YoastInputLabel } from "../YoastInput";
+import { getRtlStyle } from "../utils/helpers/styled-components";
 
 const SynonymsFieldLabelContainer = styled.span`
 	margin-bottom: 0.5em;
