+++ packages/yoast-components/composites/Plugin/Shared/components/KeywordInput.js
@@ -8,7 +8,7 @@ import noop from "lodash/noop";
 
 // Internal dependencies.
 import { colors } from "@yoast/components/style-guide";
-import { YoastInputField } from "./YoastInput";
+import { YoastInputField } from "../../../../../components/YoastInput";
 import { SvgIcon, getRtlStyle } from "@yoast/components";
 import { addFocusStyle } from "./Button";
